Python/Django/API-restaurant/requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
djangorestframework==3.13.1
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
